Inline headers from test/random
